move button and dialog directory
